Improve Hazy description of new cooling features
Clarify the intent of the changes that handle temperature floors, and
decouple them from the case of gas cooling subject to photoinization.
The operative physical mechanism is energy injection by an external
process, so generalize the discussion to include cosmic ray heating.

Seperate the discussion on the new 'stop time' command into a subsection
and discuss its use cases.