Define global YEAR constant
Previously, the code defined the constant in connection to the
'parse age' command, and used several literals with the same approximate
value across age_check.cpp.  The most commonly used value amounted to
the length of the tropical year (in seconds).

Define the constant in physconst_template.h for global access.  Use the
value of Julian year instead of the tropical (or sidereal) year, to be
consistent with the default value in astropy.  The benefit is that it is
constant, i.e., it is not subject to secular changes due to precession
or tidal effects.