Move parse_input_time into parser.cpp
Previously, the function was declared in parser.h, defined in
parse_age.cpp, and also was used in parse_stop.cpp.

Move function to parser.cpp to match the declaration (and its general
character), along with the Time namespace.  Eliminate a macro (NUMBEROF)
in favor of a constant in that namespace.

Add unit tests exercising all supported time units.