Guard against FPE with in preset time cooling sim
Previously, the code that enforced the total elapsed time preset with
'stop time' checked simply if the time would be exceed at the end of the
next time step, and adjusted it accordingly.

Peter pointed out that if the result came out to within FLT_EPSILON of
that limit, it would trigger another iteration of a time step small
enough to (likely) blow up the physics.

Relax the check on elapsed time to wrap up the calculation when the
total time was within 1e-4 of the intended total.