improve comments for ism.ini and pdr_no_dust.in
ism.ini now explains that it only turns off underabundant elements and refers to
commands to add grains or turn off elements with low abundances.

pdr_no_dust.in now explains that while it started as the Leiden F1 model it was changed to a higher
density to produce molecules in the absence of grain absorption.